+++ src/controleur/controleur_type.php
@@ -3,21 +3,29 @@
     function actionType($twig, $db){
         $form = array();
         $type = new Type($db);
+        if(isset($_POST['btSup']))
+        {
+            $ntype = $_POST['id'];
+            $type->delete($ntype);
+        }
+        if(isset($_POST['btAjout'])){
+            $libelle = $_POST['inputLibelle'];
+            $exec = $type->insert($libelle);
+        }
         $liste = $type->select();
         echo $twig->render('gestionTypes.html.twig', array('form'=>$form,'liste'=>$liste));
     }
-    function supType($twig, $db){
+    
+    function actionModifType($twig,$db){
         $form = array();
         $type = new Type($db);
-        $ntype = $_POST['id'];
-        $type->delete($ntype);
-        header('Location: index.php?page=gestionTypes');
-    }
-    function ajoutType($twig, $db){
-        $form=array();
-        $type = new Type($db);
-        $libelle = $_POST['inputLibelle'];
-        $exec = $type->insert($libelle);
-        //echo $twig->render('gestionTypes.html.twig', array('form'=>$form));
-        header("Location: index.php?page=gestionTypes");
+        $id = $_POST['id']; 
+        $unType = $type->selectByID($id);
+        if(isset($_POST['btModif'])){
+            $libelle = $_POST['inputLibelle'];
+            $type->update($id, $libelle);
+            header("Location: index.php?page=gestionTypes");
+        }
+        echo $twig->render('modifTypes.html.twig', array('type'=>$unType));
+        
     }
